+++ src/modules/Inputs/ToneGenerator.hpp
@@ -37,7 +37,7 @@ private:
 
 	bool dontUseBuffer() const override final;
 
-	bool seek(int, bool) override final;
+	bool seek(double, bool) override final;
 	bool read(Packet &, int &) override final;
 	void abort() override final;
